+++ assets/new-request-form.js
@@ -1,4 +1,4 @@
-import { j as jsxRuntimeExports, F as Field, L as Label$1, H as Hint, I as Input, M as Message, T as Textarea, a as Field$1, b as Label, c as Hint$1, C as Combobox, O as Option, d as Message$1, s as styled, A as Alert, B as Button, r as reactDomExports } from 'vendor';
+import { j as jsxRuntimeExports, F as Field, L as Label$1, H as Hint, I as Input, M as Message, T as Textarea, a as Field$1, b as Label, c as Hint$1, C as Combobox, O as Option, d as Message$1, r as reactExports, s as styled, A as Alert, B as Button, e as reactDomExports } from 'vendor';
 import { ComponentProviders } from 'shared';
 
 function TextInput({ field }) {
@@ -31,7 +31,40 @@ function TicketFormField({ label, ticketFormField, ticketForms, }) {
     return (jsxRuntimeExports.jsxs(jsxRuntimeExports.Fragment, { children: [jsxRuntimeExports.jsx("input", { type: "hidden", name: ticketFormField.name, value: ticketFormField.value }), ticketForms.length > 1 && (jsxRuntimeExports.jsxs(Field$1, { children: [jsxRuntimeExports.jsx(Label, { children: label }), jsxRuntimeExports.jsx(Combobox, { isEditable: false, onChange: handleChange, children: ticketForms.map(({ id, url, display_name }) => (jsxRuntimeExports.jsx(Option, { value: url, label: display_name, isSelected: ticketFormField.value === id, children: display_name }, id))) })] }))] }));
 }
 
-const StyledForm = styled.form `
+// NOTE: This is a temporary handling of the CSRF token
+const fetchCsrfToken = async () => {
+    const response = await fetch("/hc/api/internal/csrf_token.json");
+    const { current_session } = await response.json();
+    return current_session.csrf_token;
+};
+/**
+ * This hook creates an event handler for form submits, fetching the CSRF token
+ * from the backend and appending it to the form
+ * @returns a Submit Event Handler function
+ */
+function useSubmitHandler() {
+    const isSubmitting = reactExports.useRef(false);
+    return async (e) => {
+        e.preventDefault();
+        /* We are performing an async call to fetch the CSRF token and for this reason
+           the submit is not immediate, and the user can click the submit button multiple times.
+           We don't want to disable the submit button for A11Y, so we use the isSubmitting ref
+           to stop subsequent submits after the first one. */
+        if (isSubmitting.current === false) {
+            isSubmitting.current = true;
+            const form = e.target;
+            const token = await fetchCsrfToken();
+            const hiddenInput = document.createElement("input");
+            hiddenInput.type = "hidden";
+            hiddenInput.name = "authenticity_token";
+            hiddenInput.value = token;
+            form.appendChild(hiddenInput);
+            form.submit();
+        }
+    };
+}
+
+const Form = styled.form `
   display: flex;
   flex-direction: column;
   gap: ${(props) => props.theme.space.md};
@@ -41,26 +74,8 @@ const Footer = styled.div `
 `;
 function NewRequestForm({ ticketForms, requestForm, }) {
     const { fields, action, http_method, accept_charset, errors, ticket_form_field, ticket_forms_instructions, } = requestForm;
-    // NOTE: This is a temporary handling of the CSRF token
-    const fetchCsrfToken = async () => {
-        const response = await fetch("/hc/api/internal/csrf_token.json");
-        const { current_session } = await response.json();
-        return current_session.csrf_token;
-    };
-    const handleSubmit = (e) => {
-        e.preventDefault();
-        const form = e.target;
-        fetchCsrfToken().then((token) => {
-            const hiddenInput = document.createElement("input");
-            hiddenInput.type = "hidden";
-            hiddenInput.name = "authenticity_token";
-            hiddenInput.value = token;
-            form.appendChild(hiddenInput);
-            form.submit();
-            form.removeChild(hiddenInput);
-        });
-    };
-    return (jsxRuntimeExports.jsxs(StyledForm, { action: action, method: http_method, acceptCharset: accept_charset, noValidate: true, onSubmit: handleSubmit, children: [errors && jsxRuntimeExports.jsx(Alert, { type: "error", children: errors }), jsxRuntimeExports.jsx(TicketFormField, { label: ticket_forms_instructions, ticketFormField: ticket_form_field, ticketForms: ticketForms }), fields.map((field) => {
+    const handleSubmit = useSubmitHandler();
+    return (jsxRuntimeExports.jsxs(Form, { action: action, method: http_method, acceptCharset: accept_charset, noValidate: true, onSubmit: handleSubmit, children: [errors && jsxRuntimeExports.jsx(Alert, { type: "error", children: errors }), jsxRuntimeExports.jsx(TicketFormField, { label: ticket_forms_instructions, ticketFormField: ticket_form_field, ticketForms: ticketForms }), fields.map((field) => {
                 switch (field.type) {
                     case "anonymous_requester_email":
                     case "subject":
